+++ python/GafferUI/Viewer.py
@@ -60,6 +60,21 @@ import IECoreGL
 # without modifying the Views themselves.
 class Viewer( GafferUI.NodeSetEditor ) :
 
+	class Settings( GafferUI.Editor.Settings ) :
+
+		def __init__( self ) :
+
+			GafferUI.Editor.Settings.__init__( self )
+
+			# Receives input from the node being viewed, allowing
+			# `Editor.context()` to get the right context from the
+			# ContextTracker.
+			## \todo Can we centralise this and the behaviour of
+			# SceneEditor in NodeSetEditor?
+			self["in"] = Gaffer.Plug()
+
+	IECore.registerRunTimeTyped( Settings, typeName = "GafferUI::Viewer::Settings" )
+
 	def __init__( self, scriptNode, **kw ) :
 
 		self.__gadgetWidget = GafferUI.GadgetWidget()
@@ -194,6 +209,10 @@ class Viewer( GafferUI.NodeSetEditor ) :
 					if self.__currentView is not None :
 						break
 
+		self.settings()["in"].setInput(
+			self.__currentView["in"].getInput() if self.__currentView is not None else None
+		)
+
 		for toolbar in self.__nodeToolbars :
 			toolbar.setNode( node )
 
@@ -208,6 +227,14 @@ class Viewer( GafferUI.NodeSetEditor ) :
 			self.__updateViewportMessage()
 
 		self.__primaryToolChanged()
+
+	def _updateFromContext( self, modifiedItems ) :
+
+		if all( k.startswith( "ui:" ) for k in modifiedItems ) :
+			return
+
+		for view in self.__views :
+			view.setContext( self.context() )
 
 	def _titleFormat( self ) :