+++ Changes.md
@@ -22,6 +22,7 @@ Improvements
   - The "Double-click to toggle" tooltip is no longer displayed while hovering over non-editable cells, and a "Double-click to edit" tooltip is now displayed while hovering over other non-toggleable but editable cells.
 - Spreadsheet : Added yellow underlining to the currently active row.
 - PlugLayout : Summaries and activators are now evaluated in a context determined relative to the focus node.
+- Editor : The node graph is now evaluated in a context determined relative to the focus node.
 
 Fixes
 -----